requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
arsenic==21.8
beautifulsoup4==4.11.1
discord.py==1.7.3
firebase_admin==5.2.0
pafy==0.5.5
Pillow==9.2.0
pyppeteer==1.0.2
python-dotenv==0.20.0
PyUpload==0.1.4
selenium==4.3.0
